Trait bookworm_database::database::AppDatabase

pub trait AppDatabase {
    type Error: Send + Debug;

    fn open<'async_trait, P>(
        file_path: P
    ) -> Pin<Box<dyn Future<Output = Result<Self, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        P: AsRef<Path> + Send + Sync,
        Self: Sized,
        P: 'async_trait,
        Self: 'async_trait;
    fn path(&self) -> &Path;
    fn save<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn insert_book<'life0, 'async_trait>(
        &'life0 mut self,
        book: BookVariant
    ) -> Pin<Box<dyn Future<Output = Result<BookID, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn insert_books<'life0, 'async_trait, I>(
        &'life0 mut self,
        books: I
    ) -> Pin<Box<dyn Future<Output = Result<Vec<BookID>, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        I: 'async_trait + Iterator<Item = BookVariant> + Send,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn remove_books<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ids: &'life1 HashSet<BookID>
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn remove_selected<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        selection: &'life1 Selection
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn clear<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_book<'life0, 'async_trait>(
        &'life0 self,
        id: BookID
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Book>, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_books<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 [BookID]
    ) -> Pin<Box<dyn Future<Output = Result<HashMap<BookID, Arc<Book>>, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn read_selected_books<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        query: &'life1 str,
        bound_variables: &'life2 [Variable]
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Arc<Book>>, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn edit_book_with_id<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        id: BookID,
        edits: &'life1 [(ColumnIdentifier, Edit)]
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn edit_selected<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        selected: &'life1 Selection,
        edits: &'life2 [(ColumnIdentifier, Edit)]
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn merge_similar<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<HashSet<BookID>, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update<'life0, 'async_trait, I>(
        &'life0 mut self,
        books: I
    ) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        I: 'async_trait + Iterator<Item = BookVariant> + Send,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn has_column<'life0, 'life1, 'async_trait>(
        &'life0 self,
        col: &'life1 UniCase<String>
    ) -> Pin<Box<dyn Future<Output = Result<bool, DatabaseError<Self::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn saved<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

}

Required Associated Types

type Error: Send + Debug

Required Methods

fn open<'async_trait, P>(
file_path: P
) -> Pin<Box<dyn Future<Output = Result<Self, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
P: AsRef<Path> + Send + Sync,
Self: Sized,
P: 'async_trait,
Self: 'async_trait,

Opens the database at the path if it exists.

Arguments

file_path - A path to a database.

Errors

This function will return an error if the file points to an invalid database.

fn path(&self) -> &Path

fn save<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Saves the database to its original location.

Errors

This function will return an error if the database can not be saved correctly.

fn insert_book<'life0, 'async_trait>(
&'life0 mut self,
book: BookVariant
) -> Pin<Box<dyn Future<Output = Result<BookID, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Inserts the given book into the database, setting the ID automatically. The ID set will be returned, and calling other AppDatabase methods which take BookID with the given ID will perform functions on, or return the same book.

Arguments

book - A book to be stored.

Errors

This function will return an error if the database fails.

fn insert_books<'life0, 'async_trait, I>(
&'life0 mut self,
books: I
) -> Pin<Box<dyn Future<Output = Result<Vec<BookID>, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
I: 'async_trait + Iterator<Item = BookVariant> + Send,
'life0: 'async_trait,
Self: 'async_trait,

Stores each book into the database, and returns a Vec of corresponding IDs.

Arguments

books - Some number of books to be stored.

Errors

This function will return an error if the books can not be inserted into the database.

fn remove_books<'life0, 'life1, 'async_trait>(
&'life0 mut self,
ids: &'life1 HashSet<BookID>
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Removes all books with the given IDs. If a book with a given ID does not exist, or an ID is repeated, no changes will occur for that particular ID.

Arguments

ids - A HashSet containing the IDs of the book to be removed.

Errors

This function will return an error if the database fails.

fn remove_selected<'life0, 'life1, 'async_trait>(
&'life0 mut self,
selection: &'life1 Selection
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Removes all books contained in the provided selection.

Arguments

selection - The selection over items to be removed.

Errors

This function will return an error if the database fails.

fn clear<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_book<'life0, 'async_trait>(
&'life0 self,
id: BookID
) -> Pin<Box<dyn Future<Output = Result<Arc<Book>, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Finds and returns the book with the given ID. If no book is found, a BookNotFound error is returned.

Arguments

id - The ID of the book to be returned.

Errors

This function will return an error if the database fails or no book is found with the given ID.

fn get_books<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 [BookID]
) -> Pin<Box<dyn Future<Output = Result<HashMap<BookID, Arc<Book>>, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Finds and returns the books with the given IDs as entries in a hashmap.

Arguments

ids - The IDs of the book to be returned.

Errors

This function will return an error if the database fails

fn read_selected_books<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
query: &'life1 str,
bound_variables: &'life2 [Variable]
) -> Pin<Box<dyn Future<Output = Result<Vec<Arc<Book>>, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

fn edit_book_with_id<'life0, 'life1, 'async_trait>(
&'life0 mut self,
id: BookID,
edits: &'life1 [(ColumnIdentifier, Edit)]
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Finds the book with the given ID, then, for each pair of strings (field, new_value) in edits, set the corresponding field to new_value. If a given field is immutable, or some other failure occurs, an error will be returned.

Arguments

id - The ID of the book to be edited.

edits - A set of <field, value> pairs to set in the book.

Errors

This function will return an error if the database fails, or if any of the provided edits try to mutate an immutable column

fn edit_selected<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
selected: &'life1 Selection,
edits: &'life2 [(ColumnIdentifier, Edit)]
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Edits all books which match the provided selection, in no particular order.

Arguments

selected - a Selection over target items..

Errors

This function will return an error if the database fails, or if any of the provided edits try to mutate an immutable column

fn merge_similar<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<HashSet<BookID>, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Merges all books with matching titles and authors (case insensitive), in no particular order. Books that are merged will not necessarily free IDs no longer in use. Returns a HashSet containing the IDs of all books that have been merged.

Errors

This function will return an error if updating the database fails.

fn update<'life0, 'async_trait, I>(
&'life0 mut self,
books: I
) -> Pin<Box<dyn Future<Output = Result<(), DatabaseError<Self::Error>>> + Send + 'async_trait>> where
I: 'async_trait + Iterator<Item = BookVariant> + Send,
'life0: 'async_trait,
Self: 'async_trait,

fn has_column<'life0, 'life1, 'async_trait>(
&'life0 self,
col: &'life1 UniCase<String>
) -> Pin<Box<dyn Future<Output = Result<bool, DatabaseError<Self::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Returns whether the provided column exists in at least one book in the database.

Arguments

col - The column to check.

fn saved<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns true if the internal database is persisted to file, but does not necessarily indicate that it has been changed - eg. if a change is immediately undone, the database may still be marked as unsaved.