Trait bookworm_database::search::Matcher

pub trait Matcher: Send + Sync {
    fn new(column: ColumnIdentifier, search: String) -> Result<Self, Error>
    where
        Self: Sized;
    fn is_match(&self, book: &Book) -> bool;
    fn sql_query(&self) -> (&ColumnIdentifier, String, Option<Variable>);
    fn box_clone(&self) -> Box<dyn Matcher + Send + Sync>;
}

Provides a mechanism to determine if a particular book matches a particular search string, using some internally defined comparison method.

Required Methods

fn new(column: ColumnIdentifier, search: String) -> Result<Self, Error> where
Self: Sized,

Creates a new Matcher instance over the given search string.

fn is_match(&self, book: &Book) -> bool

Determines if the book matches the internal match rules.

fn sql_query(&self) -> (&ColumnIdentifier, String, Option<Variable>)

fn box_clone(&self) -> Box<dyn Matcher + Send + Sync>